requirements.txt
streamlit==1.18.0
google-generativeai==0.1.1
pygments==2.14.0
graphviz==0.20.1
matplotlib==3.6.3
networkx==2.8.4
python-dotenv==0.21.0

